rename file and remove duplicate fns
